Change frame rate used to record events. Refactor frame rate

+++ com.nextbreakpoint.nextfractal.core/src/main/java/com/nextbreakpoint/nextfractal/core/export/ExportHandle.java
@@ -108,7 +108,7 @@ public final class ExportHandle {
 		return session.getFrameCount();
 	}
 
-	public float getFrameRate() {
+	public int getFrameRate() {
 		return session.getFrameRate();
 	}
